creates Joystick component extracted from Channel
