+++ src/components/editor/controls/index.js
@@ -5,6 +5,7 @@ import BPMCounter from './BPMCounter/BPMCounter.jsx';
 import Channel from './Channel/Channel.jsx';
 import Controls from './TrackControls/TrackControls.jsx';
 import Dropdown from './Dropdown/Dropdown.jsx';
+import Joystick from './Joystick/Joystick.jsx';
 import PlayButton from './PlayButton/PlayButton.jsx';
 import VolumeFader from './VolumeFader/VolumeFader.jsx';
 import GlobalControls from './GlobalControls/GlobalControls.jsx';
@@ -15,6 +16,7 @@ export {
   Channel,
   Controls,
   Dropdown,
+  Joystick,
   PlayButton,
   VolumeFader,
   GlobalControls,
